+++ packages/npm-package-json-lint-config/index.js
@@ -58,6 +58,8 @@ const defaultConfig = {
 				'module',
 				'exports',
 				'react-native',
+				'wpCore',
+				'wpCoreStub',
 				'wpScript',
 				'wpScriptModuleExports',
 				'types',